requirements.txt
pysqlite3

jinja2

openai>=1.0.0
python-dotenv

beautifulsoup4
requests
faiss-cpu
tiktoken
chromadb

langgraph
langchain-core

reportlab

fastapi
uvicorn
pydantic

streamlit


# Force compatible NumPy
numpy<2.0

gunicorn

azure-communication-email

markdown


